+++ inc/debug.h
@@ -9,7 +9,7 @@
 void debug_print(const char* str);
 void debug_println(const char *format, ...);
 void debug_printf(const char *format, ...);
-void float_to_string(char *buffer, size_t size, double number);
-void debug_print_float(double num);
+void float_to_string(char *buffer, size_t size, float number);
+void debug_print_float(float num);
 
 #endif // DEBUG_H
